changed print string to show a integer instead of real number (float)

+++ CS106A-Stanford/Assignment2/hailstones.py
@@ -29,11 +29,11 @@ def hailstones_logic():
     count = 0
     while n != 1:
         if n % 2 == 0:
-            print(str(n) + ' is even, so half of it is: ' + str(n / 2))
+            print(str(n) + ' is even, so half of it is: ' + str(int(n / 2)))
             n = n / 2
             count += 1
         else:
-            print(str(n) + ' is odd, so the value of 3n + 1 is: ' + str((n * 3) + 1))
+            print(str(n) + ' is odd, so the value of 3n + 1 is: ' + str(int((n * 3) + 1)))
             n = (n * 3) + 1
             count += 1
     print('Finished... The process took ' + str(count) + ' steps to reach 1.')
